feat(revm): return journal entries
