+++ crates/primitives/src/lib.rs
@@ -4,6 +4,7 @@ pub mod bits;
 pub mod bytecode;
 pub mod db;
 pub mod env;
+pub mod journal;
 pub mod log;
 pub mod precompile;
 pub mod result;
@@ -28,7 +29,8 @@ pub use bitvec;
 pub use bytecode::*;
 pub use env::*;
 pub use hashbrown::{hash_map, HashMap};
-pub use log::Log;
+pub use journal::{Journal, JournalEntry};
+pub use log::{Log, Logs};
 pub use precompile::*;
 pub use result::*;
 pub use ruint;
